refactored + added custom edge drawing shape
