+++ README.md
@@ -16,6 +16,7 @@ The project implements a Widget for the egui framework, enabling easy visualizat
 - [x] Style configuration via egui context styles;
 - [x] Dark/Light theme support via egui context styles;
 - [x] Events reporting to extend the graph functionality by the user handling them;
+- [ ] Edge labels;
 
 ## Status
 The project is on track for a stable release v1.0.0. For the moment, breaking releases are still possible.
