+++ src/draw/mod.rs
@@ -1,5 +1,11 @@
+mod custom;
 mod drawer;
+mod edge;
 mod layers;
+mod node;
 
-pub use self::drawer::{Drawer, FnCustomEdgeDraw, FnCustomNodeDraw};
+pub use self::custom::{FnCustomEdgeDraw, FnCustomNodeDraw};
+pub use self::drawer::Drawer;
+pub use self::edge::default_edges_draw;
 pub use self::layers::Layers;
+pub use self::node::default_node_draw;
