+++ src/lib.rs
@@ -1,4 +1,5 @@
 mod computed;
+mod draw;
 mod elements;
 mod graph;
 mod graph_view;
@@ -6,9 +7,8 @@ mod metadata;
 mod settings;
 mod transform;
 
-pub mod draw;
-
 pub use self::computed::ComputedNode;
+pub use self::draw::{default_edges_draw, default_node_draw, FnCustomEdgeDraw, FnCustomNodeDraw};
 pub use self::elements::{Edge, Node};
 pub use self::graph::Graph;
 pub use self::graph_view::GraphView;
